+++ src/main/java/fi/dy/masa/malilib/mixin/MixinInGameHud.java
@@ -5,6 +5,8 @@ import net.minecraft.client.gui.DrawContext;
 import net.minecraft.client.gui.LayeredDrawer;
 import net.minecraft.client.gui.hud.InGameHud;
 import net.minecraft.client.render.RenderTickCounter;
+import net.neoforged.neoforge.client.gui.GuiLayerManager;
+import net.neoforged.neoforge.client.gui.VanillaGuiLayers;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
@@ -19,19 +21,22 @@ import fi.dy.masa.malilib.event.RenderEventHandler;
 public abstract class MixinInGameHud
 {
     @Shadow @Final private MinecraftClient client;
-    @Shadow @Final private LayeredDrawer layeredDrawer;
+    //@Shadow @Final private LayeredDrawer layeredDrawer;
+
+    @Shadow @Final private GuiLayerManager layerManager;
 
     @Inject(method = "<init>", at = @At("TAIL"))
     private void onInit(CallbackInfo info)
     {
-        this.layeredDrawer.addLayer(this::malilib_renderGameOverlayLastDrawer);
+        this.layerManager.add(VanillaGuiLayers.SUBTITLE_OVERLAY, this::malilib_renderGameOverlayLastDrawer);
+        //this.layeredDrawer.addLayer(this::malilib_renderGameOverlayLastDrawer);
     }
 
-//    @Inject(method = "render", at = @At("TAIL"))
-//    private void malilib_onGameOverlayPost(DrawContext context, RenderTickCounter tickCounter, CallbackInfo ci)
-//    {
-//        ((RenderEventHandler) RenderEventHandler.getInstance()).onRenderGameOverlayPost(context, this.client, tickCounter.getTickDelta(false));
-//    }
+    @Inject(method = "render", at = @At("TAIL"))
+    private void malilib_onGameOverlayPost(DrawContext context, RenderTickCounter tickCounter, CallbackInfo ci)
+    {
+        ((RenderEventHandler) RenderEventHandler.getInstance()).onRenderGameOverlayPost(context, this.client, tickCounter.getTickDelta(false));
+    }
 
     @Unique
     private void malilib_renderGameOverlayLastDrawer(DrawContext context, RenderTickCounter tickCounter)