+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 ## Changelog
 - Use our own modification of FFAPI
+- fix render duplicate HUD